# arcanium/api/openapi/arcanium.yaml
openapi: 3.1.0
info:
  title: Arcanium API
  version: "1.0.0"
  description: >
    Arcanium's cryptographic control-plane API. Authored from the actual
    Express route handlers (arcanium/api/src/routes/*.js,
    arcanium/api/src/auth/index.js) as source of truth for current
    behavior — a capture-then-enforce exercise (Prompt 22), not a redesign.
    This is the ONE authoritative contract: the Nuxt gateway allowlist
    (arcanium/ui/server/routes/gateway/[...path].ts), the CLI request
    shapes, and docs/api.md's narrative must all be checked against this
    file, not maintained as independent parallel sources of truth.
servers:
  - url: http://localhost:3001
    description: Direct API (dev/scenario scripts)
  - url: http://localhost:3000/gateway
    description: Same-origin UI gateway (browser-facing; allowlisted subset only)
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: arc_session
      description: >
        Opaque session id issued by GET /api/v1/auth/callback. When
        ARCANIUM_AUTH_ENABLED is unset, requireSession no-ops and every
        request is treated as an estate-wide operator — see
        docs/security.md.
    serviceAccountBearer:
      type: http
      scheme: bearer
      description: >
        Reserved for a future machine-to-machine identity source
        (prompts/26_external_integration_key_lifecycle.md, not yet
        executed). Not accepted by any route in this version of the API —
        listed here only so the contract does not need a breaking change
        once it is.
  responses:
    Unauthorized:
      description: No session cookie, or the session is expired/invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >
        Authenticated, but authorize() denied the action — wrong role,
        wrong tenant, or a scoped grant that did not match (Phase 18's
        deny-by-default matrix, auth/authorize.js).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: >
        Missing, or deliberately indistinguishable from missing for a
        cross-tenant read (Phase 18's GET /applications/:id precedent —
        a supplier-admin gets 404, not 403, for another tenant's resource,
        so existence itself isn't leaked).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Request failed field validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Direct-API responses use a string; gateway responses use boolean true.
          oneOf:
            - type: string
            - type: boolean
        statusCode: { type: integer }
        statusMessage: { type: string }
        message: { type: string }
        field: { type: string, description: "Direct-API 400 responses name the offending field." }
        reason: { type: string, description: "authorize() denial reason, on 403 responses." }
        action: { type: string, description: "The authorize() action that was denied, on 403 responses." }
      required: [error]

    # ── Suppliers / Applications ─────────────────────────────────────────
...
    Team:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: platform }
        description: { type: ["string", "null"] }
        supplier_ids:
          type: ["array", "null"]
          items: { type: string, format: uuid }
          description: "null = every supplier (an estate-wide-by-team audit grant)"
        environments:
          type: ["array", "null"]
          items: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, created_by, created_at]

    # ── Transit keys ──────────────────────────────────────────────────────
...
paths:
...
  /api/v1/teams:
    get:
      summary: List teams (estate-wide read only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Team" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Register a team (links a "arcanium-<role>:team:<name>" OIDC group to supplier ids) — estate-wide provision only
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: ["string", "null"] }
                supplier_ids: { type: ["array", "null"], items: { type: string, format: uuid } }
                environments: { type: ["array", "null"], items: { type: string } }
              required: [name]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Team name already registered }
  /api/v1/teams/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Team detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update a team's description/supplier_ids/environments — estate-wide provision only
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Remove a team — estate-wide provision only
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Approvals ─────────────────────────────────────────────────────────
...